Struct steap_by_steap::rwaytrie::RWayTrie

pub struct RWayTrie {
    root: Option<Box<Node>>,
}

An R-way Trie data structure.

Fields
root: Option<Box<Node>>
Implementations

impl RWayTrie

fn new() -> Self

Create a new empty R-way Trie.

fn insert(&mut self, key: &str, value: &str)

Insert a key-value pair into the trie.

Args: key (str): The key to insert. value (str): The value associated with the key.

fn search(&self, key: &str) -> Option<String>

Search for a key in the trie.

Args: key (str): The key to search for.

Returns: Optionalstr: The value associated with the key, or None if not found.

fn keys_with_prefix(&self, prefix: &str) -> PyResult<HashMap<String, String>>

Get all key-value pairs in the trie with a given prefix.

Args: prefix (str): The prefix to search for.

Returns: Dict[str, str]: A dictionary of key-value pairs with the given prefix.

impl RWayTrie

fn collect( &self, node: &Node, prefix: &mut String, results: &mut HashMap<String, String> )