Struct steap_by_steap::weightedgraph::WeightedGraph

pub struct WeightedGraph {
    adjacency_list: HashMap<i32, Vec<(i32, i32)>>,
}

Fields
adjacency_list: HashMap<i32, Vec<(i32, i32)>>
Implementations

impl WeightedGraph

fn new() -> Self

fn add_edge(&mut self, from_node: i32, to_node: i32, weight: i32)

Add an edge to the graph

Args: from_node (int): The source node to_node (int): The destination node weight (int): The weight of the edge

fn dijkstra(&self, start: i32, end: i32) -> PyResult<(i32, Vec<i32>)>

Find the shortest path using Dijkstra’s algorithm

Args: start (int): The starting node end (int): The ending node

Returns: tuple: A tuple containing the total cost and the path (as a list of nodes)

fn bellman_ford(&self, start: i32, end: i32) -> PyResult<(i32, Vec<i32>)>

Find the shortest path using Bellman-Ford algorithm

Args: start (int): The starting node end (int): The ending node

Returns: tuple: A tuple containing the total cost and the path (as a list of nodes)

fn prim(&self) -> PyResult<Vec<(i32, i32, i32)>>

Find the Minimum Spanning Tree using Prim’s algorithm

Returns: list: A list of tuples representing the edges in the MST (from_node, to_node, weight)

fn kruskal(&self) -> PyResult<Vec<(i32, i32, i32)>>

Find the Minimum Spanning Tree using Kruskal’s algorithm

Returns: list: A list of tuples representing the edges in the MST (from_node, to_node, weight)

fn ford_fulkerson(&mut self, source: i32, sink: i32) -> PyResult<i32>

Find the maximum flow using Ford-Fulkerson algorithm

Args: source (int): The source node sink (int): The sink node

Returns: int: The maximum flow from source to sink

fn bfs_find_path(&self, source: i32, sink: i32) -> (Vec<i32>, i32)